[OC-1440] New logging screen filters
