+++ ui/main/src/app/modules/archives/archives.component.html
@@ -12,61 +12,7 @@
       <div class="opfab-archives">
         <div style="display: flex;">
 
-          <!--2 columns in this case-->
-          <!--First column-->
-          <div *ngIf="displayServiceFilter() || tags" style="margin-right:40px">
-            <div  style="min-width:300px;max-width:400px;margin-top: 28px;margin-bottom: 54px;">
-              <of-multi-filter *ngIf="displayServiceFilter()" filterPath="service" id="service" [parentForm]="archiveForm"
-                               [values]="serviceDropdownList" [dropdownSettings]="serviceDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-              <of-multi-filter *ngIf="! displayServiceFilter() && tags" filterPath="tags" id="tags" [parentForm]="archiveForm"
-                               [values]="tagsDropdownList" [dropdownSettings]="tagsDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-            <div *ngIf="displayServiceFilter() && tags" style="min-width:300px;max-width:400px;">
-              <of-multi-filter filterPath="tags" id="tags" [parentForm]="archiveForm" [values]="tagsDropdownList"
-                               [dropdownSettings]="tagsDropdownSettings" [selectedItems]="[]" i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-          </div>
-
-          <!--Second column-->
-          <div *ngIf="displayServiceFilter() || tags" style="margin-right:40px">
-            <div style="min-width:300px;max-width:400px;margin-top:28px;margin-bottom: 54px;">
-              <of-multi-filter [hidden]="displayServiceFilter()" filterPath="process" id="process" [parentForm]="archiveForm"
-                               [values]="processDropdownList" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-              <of-multi-filter [hidden]="! displayServiceFilter()" filterPath="process" id="process" [parentForm]="archiveForm"
-                               [values]="processDropdownListWhenSelectedService" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-            <div style="min-width:300px;max-width:400px;">
-              <of-multi-filter filterPath="state" id="state" [parentForm]="archiveForm"
-                               [values]="stateDropdownListWhenSelectedProcess" [dropdownSettings]="stateDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-          </div>
-
-          <!--Only one column in this case-->
-          <div *ngIf="! displayServiceFilter() && ! tags" style="margin-right:40px">
-            <div style="min-width:300px;max-width:400px;margin-top:28px;margin-bottom: 54px;">
-              <of-multi-filter filterPath="process" id="process" [parentForm]="archiveForm"
-                               [values]="processDropdownList" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-            <div style="min-width:300px;max-width:400px;">
-              <of-multi-filter filterPath="state" id="state" [parentForm]="archiveForm"
-                               [values]="stateDropdownListWhenSelectedProcess" [dropdownSettings]="stateDropdownSettings" [selectedItems]="[]"
-                               i18nRootLabelKey="archive.filters.">
-              </of-multi-filter>
-            </div>
-          </div>
+          <of-archives-logging-filters [parentForm]="archiveForm" #filters></of-archives-logging-filters>
 
           <div class="opfab-vertical-bar"></div>
           <div style="margin-left:40px;margin-right: 30px;width:230px; min-width: 230px;">
@@ -113,7 +59,7 @@
           <th scope="col" translate>archive.result.businessPeriod</th>
           <th scope="col" translate>archive.result.title</th>
           <th scope="col" translate>archive.result.summary</th>
-          <th *ngIf="displayServiceFilter()" scope="col" translate>archive.result.service</th>
+          <th *ngIf="filtersTemplate.displayServiceFilter()" scope="col" translate>archive.result.service</th>
         </tr>
       </thead>
 
@@ -127,7 +73,7 @@
         <td translate [translateParams]="card.summary.parameters">
           {{ card.process + '.' + card.processVersion + '.' + card.summary.key}}
         </td>
-        <td *ngIf="displayServiceFilter()" translate>{{ findServiceLabelForProcess(card.process) }}</td>
+        <td *ngIf="filtersTemplate.displayServiceFilter()" translate>{{ filtersTemplate.findServiceLabelForProcess(card.process) }}</td>
       </tr>
     </table>
     <div class="opfab-archive-pagination">
