+++ ui/main/src/app/modules/logging/components/logging-filters/logging-filters.component.html
@@ -5,19 +5,13 @@
 <!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
 <!-- SPDX-License-Identifier: MPL-2.0                                      -->
 <!-- This file is part of the OperatorFabric project.                      -->
+
 <form [formGroup]="loggingForm" #currentForm>
     <div class="opfab-logging" >
         <div style="display: flex;">
-            <div  style="margin-top: 28px;margin-right: 40px;position:relative; min-width:300px;max-width:400px;" >
-                <of-multi-filter filterPath="process"
-                                   id="process"
-                                   [parentForm]="loggingForm"
-                                   [values]="dropdownList"
-                                   [dropdownSettings]="dropdownSettings"
-                                   [selectedItems]="[]"
-                                   i18nRootLabelKey="logging.filters."
-                ></of-multi-filter>
-            </div>
+
+            <of-archives-logging-filters [parentForm]="loggingForm" #filters></of-archives-logging-filters>
+
             <div class="opfab-vertical-bar"></div>
             <div style="margin-left:40px;margin-right: 30px;width:230px;min-width: 230px;">
                 <div>
