+++ ui/main/src/app/modules/logging/components/logging-filters/logging-filters.component.ts
@@ -7,7 +7,7 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {Component, Input, OnDestroy, OnInit} from '@angular/core';
+import {Component, Input, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {AppState} from '@ofStore/index';
 import {Store} from '@ngrx/store';
 import {FormControl, FormGroup} from '@angular/forms';
@@ -15,34 +15,9 @@ import {FormControl, FormGroup} from '@angular/forms';
 import {SendLoggingQuery} from '@ofActions/logging.actions';
 import {ConfigService} from '@ofServices/config.service';
 import {TimeService} from '@ofServices/time.service';
-import {NgbDateStruct, NgbTimeStruct} from '@ng-bootstrap/ng-bootstrap';
-import {DateTimeNgb} from '@ofModel/datetime-ngb.model';
-import {TranslateService} from '@ngx-translate/core';
-import {buildSettingsOrConfigSelector} from '@ofStore/selectors/settings.x.config.selectors';
-import {Observable, Subject} from 'rxjs';
-import {takeUntil} from 'rxjs/operators';
+import {Subject} from 'rxjs';
+import {ArchivesLoggingFiltersComponent} from "../../../share/archives-logging-filters/archives-logging-filters.component";
 
-
-export enum FilterDateTypes {
-    PUBLISH_DATE_FROM_PARAM = 'publishDateFrom',
-    PUBLISH_DATE_TO_PARAM = 'publishDateTo',
-    ACTIVE_FROM_PARAM = 'activeFrom',
-    ACTIVE_TO_PARAM = 'activeTo'
-
-}
-
-
-export const checkElement = (enumeration: typeof FilterDateTypes, value: string): boolean => {
-    let result = false;
-    if (Object.values(enumeration).map(enumValue => enumValue.toString()).includes(value)) {
-        result = true;
-    }
-    return result;
-};
-
-export const transformToTimestamp = (date: NgbDateStruct, time: NgbTimeStruct): string => {
-    return new DateTimeNgb(date, time).formatDateTime();
-};
 
 @Component({
     selector: 'of-logging-filters',
@@ -57,15 +32,16 @@ export class LoggingFiltersComponent implements OnInit, OnDestroy {
     size = 10;
     loggingForm: FormGroup;
 
-    dropdownList = [];
-    dropdownSettings = {};
+    @ViewChild('filters') filtersTemplate: ArchivesLoggingFiltersComponent;
 
     public submittedOnce = false;
 
     @Input()
     public processData: [];
 
-    constructor(private store: Store<AppState>, private timeService: TimeService, private configService: ConfigService, private translate: TranslateService) {
+    constructor(private store: Store<AppState>,
+                private timeService: TimeService,
+                private configService: ConfigService) {
 
     }
 
@@ -73,71 +49,26 @@ export class LoggingFiltersComponent implements OnInit, OnDestroy {
         this.size = this.configService.getConfigValue('archive.filters.page.size', 10);
         this.loggingForm = new FormGroup(
             {
+                tags: new FormControl([]),
+                state: new FormControl([]),
                 process: new FormControl([]),
+                service: new FormControl([]),
                 publishDateFrom: new FormControl(''),
                 publishDateTo: new FormControl(''),
                 activeFrom: new FormControl(''),
                 activeTo: new FormControl('')
             }
         );
-        this.dropdownList = this.processData;
-
-        this.getLocale().pipe(takeUntil(this.unsubscribe$)).subscribe(locale => {
-            this.translate.use(locale);
-            this.translate.get(['logging.filters.selectProcessText'])
-              .subscribe(translations => {
-                this.dropdownSettings = {
-                    text: translations['logging.filters.selectProcessText'],
-                    enableSearchFilter: true,
-                    badgeShowLimit: 4,
-                    classes: 'custom-class-example'
-                }
-              })
-            });
-    }
-
-    protected getLocale(): Observable<string> {
-        return this.store.select(buildSettingsOrConfigSelector('locale'));
     }
 
     sendQuery() {
         const {value} = this.loggingForm;
-        const params = this.filtersToMap(value);
-        params.set('size', [this.size.toString()]);
-        params.set('page', ['0']);
-        this.store.dispatch(new SendLoggingQuery({params}));
+        this.filtersTemplate.filtersToMap(value);
+        this.filtersTemplate.filters.set('size', [this.size.toString()]);
+        this.filtersTemplate.filters.set('page', ['0']);
+        this.store.dispatch(new SendLoggingQuery({params: this.filtersTemplate.filters}));
         this.submittedOnce = true;
     }
-
-    /**
-     * Transforms the filters list to Map
-     */
-    filtersToMap = (filters: any): Map<string, string[]> => {
-        const params = new Map();
-        Object.keys(filters).forEach(key => {
-            const element = filters[key];
-            // if the form element is date
-            if (element) {
-                if (checkElement(FilterDateTypes, key)) {
-                    const {date, time} = element;
-                    if (date) {
-                        const timeStamp = this.timeService.toNgBTimestamp(transformToTimestamp(date, time));
-                        if (timeStamp !== 'NaN') {
-                            params.set(key, [timeStamp]);
-                        }
-                    }
-                } else {
-                    if (element.length) {
-                        const idProcessArray = [];
-                        element.forEach(val => idProcessArray.push(val.id));
-                        params.set(key, idProcessArray);
-                    }
-                }
-            }
-        });
-        return params;
-    }
-
 
     resetForm()
     {
