+++ ui/main/src/app/modules/logging/logging.component.html
@@ -15,7 +15,10 @@
 </div>
 
 <div  *ngIf="(loggingResult$|async) as line ; else noResult" style="margin-left:5%;margin-right:5%">
-    <of-logging-table [results]="line" [processStateDescription]="processStateDescription"></of-logging-table>
+    <of-logging-table [results]="line"
+                      [processStateDescription]="processStateDescription"
+                      [displayServiceColumn]="this.filters.filtersTemplate.displayServiceFilter()" >
+    </of-logging-table>
 </div>
 
 <ng-template #noResult>
