+++ ui/main/src/app/modules/share/archives-logging-filters/archives-logging-filters.component.html
@@ -1,0 +1,65 @@
+<!-- Copyright (c) 2021, RTE (http://www.rte-france.com)              -->
+<!-- See AUTHORS.txt                                                       -->
+<!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
+<!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
+<!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
+<!-- SPDX-License-Identifier: MPL-2.0                                      -->
+<!-- This file is part of the OperatorFabric project.                      -->
+
+<div style="display: flex;">
+<!--2 columns in this case-->
+<!--First column-->
+<div *ngIf="displayServiceFilter() || tags" style="margin-right:40px">
+    <div  style="min-width:300px;max-width:400px;margin-top: 28px;margin-bottom: 54px;">
+        <of-multi-filter *ngIf="displayServiceFilter()" filterPath="service" id="service" [parentForm]="this.parentForm"
+                         [values]="serviceDropdownList" [dropdownSettings]="serviceDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+        <of-multi-filter *ngIf="! displayServiceFilter() && tags" filterPath="tags" id="tags" [parentForm]="this.parentForm"
+                         [values]="tagsDropdownList" [dropdownSettings]="tagsDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+    <div *ngIf="displayServiceFilter() && tags" style="min-width:300px;max-width:400px;">
+        <of-multi-filter filterPath="tags" id="tags" [parentForm]="this.parentForm" [values]="tagsDropdownList"
+                         [dropdownSettings]="tagsDropdownSettings" [selectedItems]="[]" i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+</div>
+
+<!--Second column-->
+<div *ngIf="displayServiceFilter() || tags" style="margin-right:40px">
+    <div style="min-width:300px;max-width:400px;margin-top:28px;margin-bottom: 54px;">
+        <of-multi-filter [hidden]="displayServiceFilter()" filterPath="process" id="process" [parentForm]="this.parentForm"
+                         [values]="processDropdownList" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+        <of-multi-filter [hidden]="! displayServiceFilter()" filterPath="process" id="process" [parentForm]="this.parentForm"
+                         [values]="processDropdownListWhenSelectedService" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+    <div style="min-width:300px;max-width:400px;">
+        <of-multi-filter filterPath="state" id="state" [parentForm]="this.parentForm"
+                         [values]="stateDropdownListWhenSelectedProcess" [dropdownSettings]="stateDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+</div>
+
+<!--Only one column in this case-->
+<div *ngIf="! displayServiceFilter() && ! tags" style="margin-right:40px">
+    <div style="min-width:300px;max-width:400px;margin-top:28px;margin-bottom: 54px;">
+        <of-multi-filter filterPath="process" id="process" [parentForm]="this.parentForm"
+                         [values]="processDropdownList" [dropdownSettings]="processDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+    <div style="min-width:300px;max-width:400px;">
+        <of-multi-filter filterPath="state" id="state" [parentForm]="this.parentForm"
+                         [values]="stateDropdownListWhenSelectedProcess" [dropdownSettings]="stateDropdownSettings" [selectedItems]="[]"
+                         i18nRootLabelKey="filters.">
+        </of-multi-filter>
+    </div>
+</div>
+</div>
